+++ include/seom/config.h
@@ -5,6 +5,6 @@
 void seomConfigServer(const char *ns, char server[256]);
 void seomConfigInterval(const char *ns, double *v);
 void seomConfigScale(const char *ns, char scale[64]);
-void seomConfigInsets(const char *ns, uint64_t v[4]);
+void seomConfigInsets(const char *ns, uint32_t v[4]);
 
 #endif /* __SEOM_CONFIG_H__ */
